RTI does not change B and U flags
The obelisk reference states that the break flag is set from stack, but
the nesdev reference states that the break flag is not set from stack. I
will assume the NES chip behaves slightly differently from the 6502 and
does not change flags 4 & 5 in the status register. May need to revisit
if they are set to default values rather than left alone.