Make processor pass all SingeStepTest/65x02 -> nes6502/v1/ tests
Using a separate project designed to run and test the cpu in known
states from the SingleStepTest/65x02 repo, I've fixed the Indirect X/Y
addressing as well as a few bugs in detecting accumulator addressing in
RWM instructions. The ROR instruction was also missing the shift
operator lol. It also turns out that the PHP always pushed the U and B
flags, even though it is not a branch instruction.