Begin moving to half-cycle accurate emulation
I've always wanted to do this, and I belive this was the best time to
bite the bullet and actually *do* it. This commit brings the emualtor
closer to being half-cycle accurate where the bus is not a value of the
cpu and instead data is read/written to bus 'pins' based on the phi
signals. This will enable better high-level control of the bus later and
breaks down the instructions into even simpler units of code. I hope
this eventually turns into something akin to a microcode.