Clean up, formatting, and organizing debugging logic
Cleaned up some unnecessary lines and variables.
Formatted the conditions to look nicer in sequence.
Due to move of non-copyable data outside the cpu, the cpu is now
copy-able and can be copied by the debug handler at the end of an
instruction. The new architecture should also allow for grabbing debug
information without interfering with the cpu's execution.
Also optimized printing by using a pre-sized buffer and manual
"formatting".