Move binary application to executable
The winit+wpgu+egui was a triple-combo on dependencies and compile time.
Moving the windowing features to an example reduces the total
dependencies to a grand-total of: 1. Bitflags, that's it.
If I ever implement an official status+code viewer, I will probably use
a TUI library or a separate project.