Cpu doesn't know how many cycles it has run for, more Assembly generation
Remove cycles from being inside of the cpu, the cpu does not know how
long it has been running.
Add the unofficial opcodes to the name mapping function.
Determine if an opcode is unofficial by appending an astrik before the
name in the debug log.
Comment out more unused code.