Use structs to represent pin IO, begin PPU logic and rendering
Cpu clock functions had too many arguments, all related to pin IO. Moved
this data into a struct, followed suit for the Ppu.
Began Ppu logic, simply fills a buffer with a magenta rgba color and
draws it to the screen on the NesTest example - will make a suitable
example for rendering later.
Run the cpu at about the speed required for a full Ppu frame. Its
missing the VBlank period in the cycle count, but that will be fixed
when a new example is made.