Clean up Option<T> spam from the Application
Due to the nature of winit's new Application design, a window is not
available until after the event loop is active (and only accessible
within the Application's member methods). Due to this, it's not possible
to create many member variables that would normally live for the
entirety of the Application until after the Application is made.
This commit create a new wrapping Application that manages the state of
being Uninitialized, Resumed, and Suspended that wraps around the old
App struct.
This change cleans up many Optional unwrapping and destructuring.