Scaled up video output, Moved debug images to a collapsable, added ui buttons for cpu interrupt, deleted some unused code
Moved the images drawn around the game screen into the side panel under
a collapsable menu. This also ensures that generating the nametable is
only done on frames where it is visible.
Since clutter was moved away, the rendered texture of the actual output
could be scaled up 2x.
Removed the keybinds for the cpu interrupts and replaced them with ui
buttons in the side panel.
Cleaned up the code related to image rendering.
Deleted some code related to the old ram rendering.