Fix sprite rendering issues (and the freeze when touching the SMB sprite-0 coin) and clean up code a little
For the longest time, SMB would freeze if any sprite overlapped with
sprite-0 (the scoreboard coin used to detect when the scoreboard should
end and scrolling be enabled) and would cause sprite-0 to be overridden.
This was caused because the OAM memory was advanced by 4 every cycle of
the evaluation instead of only the last cycle, causing the evaluation to
start at sprite 3 @ address 12 instead of 0. Infact, sprite-0 was being
evaluated in the bottom 1/3 of sprites meaning 2/3 of sprites had higher
priority and would fill the secondary OAM before it.